Move ObjectFactory to core module
