+++ java/compaction/compaction-job-execution/src/main/java/sleeper/compaction/job/execution/DefaultCompactionRunnerFactory.java
@@ -23,9 +23,9 @@ import sleeper.compaction.core.job.CompactionJob;
 import sleeper.compaction.core.job.CompactionRunner;
 import sleeper.compaction.core.task.CompactionRunnerFactory;
 import sleeper.compaction.rust.RustCompactionRunner;
-import sleeper.configuration.jars.ObjectFactory;
 import sleeper.core.properties.table.TableProperties;
 import sleeper.core.properties.validation.CompactionMethod;
+import sleeper.core.util.ObjectFactory;
 
 import static sleeper.core.properties.table.TableProperty.COMPACTION_METHOD;
