+++ java/configuration/src/test/java/sleeper/configuration/jars/S3UserJarsLoaderIT.java
@@ -29,6 +29,8 @@ import org.testcontainers.utility.DockerImageName;
 import sleeper.core.CommonTestConstants;
 import sleeper.core.iterator.SortedRecordIterator;
 import sleeper.core.properties.instance.InstanceProperties;
+import sleeper.core.util.ObjectFactory;
+import sleeper.core.util.ObjectFactoryException;
 
 import javax.tools.JavaFileObject;
 import javax.tools.SimpleJavaFileObject;
