+++ java/core/src/main/java/sleeper/core/util/ObjectFactory.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.configuration.jars;
+package sleeper.core.util;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
