+++ java/system-test/system-test-drivers/src/main/java/sleeper/systemtest/drivers/ingest/AwsDirectIngestDriver.java
@@ -18,10 +18,10 @@ package sleeper.systemtest.drivers.ingest;
 
 import software.amazon.awssdk.services.s3.S3AsyncClient;
 
-import sleeper.configuration.jars.ObjectFactory;
 import sleeper.core.iterator.IteratorCreationException;
 import sleeper.core.record.Record;
 import sleeper.core.statestore.StateStoreException;
+import sleeper.core.util.ObjectFactory;
 import sleeper.ingest.runner.IngestFactory;
 import sleeper.systemtest.drivers.util.SystemTestClients;
 import sleeper.systemtest.dsl.ingest.DirectIngestDriver;
